add ContactHelper, NavigationHelper and GroupHelper
